% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_half_moons}
\alias{two_half_moons}
\title{Two interlocking half moons data set}
\format{
An object of class \code{data.frame} with 150 rows and 3 columns.
}
\usage{
data(two_half_moons)
}
\description{
A dataset containing 150 observations generated according to the
two interlocking half moons data generating process. The first two columns
contain the x and y-coordinates and the third column contains the cluster ID.
Each moon contains 75 observations.
}
\keyword{datasets}
